feat: Atualizando regra de negócio para remover comentário

+++ Core/BlogMVC.Application/Services/CommentService.cs
@@ -13,17 +13,20 @@ public class CommentService : ICommentService
     private readonly IPostRepository _postRepository;
     private readonly IUserRepository _userRepository;
     private readonly IUserLogged _userLogged;
+    private readonly IUnityOfWork _unityOfWork;
     private readonly IMapper _mapper;
     public CommentService(ICommentRepository commentRepository, 
         IPostRepository postRepository, 
         IMapper mapper, IUserLogged userLogged, 
-        IUserRepository userRepository)
+        IUserRepository userRepository, 
+        IUnityOfWork unityOfWork)
     {
         _commentRepository = commentRepository; 
         _postRepository = postRepository;
         _mapper = mapper; 
         _userLogged = userLogged;
         _userRepository = userRepository;
+        _unityOfWork = unityOfWork; 
     }
     public async Task<GetCommentsResponseJson> Create(CreateCommentRequestJson request, int postId)
     {
@@ -57,11 +60,13 @@ public class CommentService : ICommentService
 
     public async Task<bool> Remove(int commentId)
     {
+        var user = await _userLogged.GetUser(); 
         var comment = await _commentRepository.GetByIdAsync(commentId); 
-        if(comment is null)
-            throw new BlogException("Comentário não encontrado"); 
+        if((comment is null) || comment.UserId != user.Id)
+            throw new NotFoundException(); 
         
         await _commentRepository.RemoveAsync(comment); 
+        await _unityOfWork.Commit(); 
         return true; 
     }
